Rename DTK_Details{Point,Box,Sphere}.hpp => DTK_{Point,Box,Sphere}.hpp
