+++ packages/Search/src/details/DTK_DetailsTeuchosSerializationTraits.hpp
@@ -11,10 +11,10 @@
 #ifndef DTK_DETAILS_TEUCHOS_SERIALIZATION_TRAITS_HPP
 #define DTK_DETAILS_TEUCHOS_SERIALIZATION_TRAITS_HPP
 
-#include <DTK_DetailsBox.hpp>
-#include <DTK_DetailsPoint.hpp>
+#include <DTK_Box.hpp>
 #include <DTK_DetailsPredicate.hpp>
-#include <DTK_DetailsSphere.hpp>
+#include <DTK_Point.hpp>
+#include <DTK_Sphere.hpp>
 
 #include <Teuchos_SerializationTraits.hpp>
