+++ packages/Search/test/DTK_BoostRTreeHelpers.hpp
@@ -12,9 +12,9 @@
 #ifndef DTK_BOOST_RTREE_HELPERS_HPP
 #define DTK_BOOST_RTREE_HELPERS_HPP
 
-#include <DTK_DetailsBox.hpp>
-#include <DTK_DetailsPoint.hpp>
+#include <DTK_Box.hpp>
 #include <DTK_DetailsPredicate.hpp>
+#include <DTK_Point.hpp>
 
 #include "DTK_BoostGeometryAdapters.hpp"
 #include "DTK_BoostRangeAdapters.hpp"
